requirements.txt
# Core
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.30.0,<1.0
pydantic>=2.7.0,<3.0
pydantic-settings>=2.3.0,<3.0

# Anthropic AI
anthropic>=0.39.0,<1.0

# Kubernetes
kubernetes>=31.0.0,<36.0

# PostgreSQL
asyncpg>=0.29.0,<1.0
pgvector>=0.3.0,<1.0

# Embeddings (Voyage AI — uses Anthropic API key)
voyageai>=0.3.0,<1.0

# Slack
slack-sdk>=3.31.0,<4.0

# YAML (for runbooks)
pyyaml>=6.0,<7.0

# Testing (dev only)
# pytest>=8.0
# pytest-asyncio>=0.24.0
# httpx>=0.27.0
# respx>=0.21.0
